Finish booking activity page. Redesign route path in reservation page.

+++ src/components/checkout/SidebarRoom/SidebarRoom.vue
@@ -1,20 +1,20 @@
 <template>
   <div>
-    <div>
+    <div class="room__content" :class="onSelectClass">
       <div class="flex--column">
-        <div class="flex--row side__content--row">
+        <div class="flex--row room__content--row">
           <span><u><i>{{ info.roomSelect.roomType }}</i></u></span>
           <p></p>
         </div>
-        <div class="flex--row side__content--row">
+        <div class="flex--row room__content--row">
           <h5>{{ startDate }} - {{ endDate }}</h5>
           <p></p>
         </div>
-        <div class="flex--row side__content--row">
+        <div class="flex--row room__content--row">
           <h5><i>{{ info.totalNight }} night(s)</i></h5>
           <p></p>
         </div>
-        <div class="flex--row side__content--row">
+        <div class="flex--row room__content--row">
           <h5>{{ info.guests.numOfAdultGuests }} Adults, {{ info.guests.numOfChildrenGuest }} Children</h5>
           <p></p>
         </div>
@@ -29,11 +29,11 @@
           </li>
         </ul>
       </div>
-      <div class="flex--row side__content--row">
+      <div class="flex--row room__content--row">
         <span>Taxes and Fees</span>
         <p>${{ calculateTax }}</p>
       </div>
-      <div class="flex--row side__content--action">
+      <div class="flex--row room__content--action">
         <span @click="editRoom">Edit</span> |
         <span @click="removeRoom">Remove</span>
       </div>
@@ -42,19 +42,29 @@
 </template>
 <script>
 import dayjs from 'dayjs'
-
+import { mapState } from 'vuex'
 export default {
   props: {
     reservationInfo: {
       type: Object,
       required: true
+    },
+    id: {
+      type: Number,
+      required: true
     }
   },
   data () {
     return {
+      dialogVisible: false
     }
   },
   computed: {
+    ...mapState({
+      tempId: state => state.reservation.tempId,
+      time: state => state.reservation.reservationDetails.createTime,
+      onSelectIdx: state => state.reservation.currentSelectedRoomIdx
+    }),
     calculateTax () {
       return Math.round((this.info.roomSelect.rate + this.calculateAddOnPrice) * 0.08 * 100) / 100
     },
@@ -71,16 +81,44 @@ export default {
     },
     endDate () {
       return dayjs(this.info.date.end).format('ddd MM/DD/YYYY')
+    },
+    onSelectClass () {
+      return this.id === this.onSelectIdx ? 'selected-class' : null
     }
   },
   methods: {
     removeRoom () {
-      this.$store.dispatch('removeRoomFromSelection', this.info.createTime)
-      this.$store.dispatch('calculateTotalAmount')
+      this.$confirm('Are you sure you want to delete this selection? Select YES to confirm selection removal.',
+        'Confirmation',
+        {
+          confirmButtonText: 'YES',
+          cancelButtonText: 'CANCEL',
+          type: 'warning'
+        }).then(() => {
+        this.$store.dispatch('removeRoomFromSelection', this.info.createTime)
+        this.$store.dispatch('calculateTotalAmount')
+        this.$message({
+          type: 'success',
+          message: 'Remove completed'
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: 'Remove canceled'
+        })
+      })
     },
     editRoom () {
       this.$store.dispatch('editRoomFromSelection', this.info.createTime)
-      this.$router.push('/reservation/s1')
+      this.$router.push({
+        name: 'Reservation',
+        params: { tempId: this.tempId },
+        query: {
+          createdTime: dayjs(this.time).format('YYYY-MM-DD'),
+          currentStep: 's1',
+          prevStep: 's3'
+        }
+      })
     }
   }
 }
